Contains method has been changed to getIndex
It now fetches the index of a character and throws an Exception if no character is found. Also work has been done on the FrequencyTable(String input) constructor to read the input, but it does not count yet. A stub for incrementing a character that we already have has been written